requirements.txt
dimcli==1.1
matplotlib==3.7.4
pandas==2.0.3
python-dotenv==1.0.0